feat(Query): renamed Query to AbstractQuery, added in concrete Query
In what feels like the final turn of thinking on the whitepaper, I've replicated
the concrete / abstract pattern for queries. This arose while working through
how to restore query deduplication now that metadata hashes shift with timestamps,
which made it clear that we should be recording a hash that combines an abstract
query definition with the hashes of the resources it's run against.

As more proof this is a step in the right direction, once Query & AbstractQuery
both existed, it became clear that a number of fields in Dataset prefixed
with 'Query' could be moved into this new concrete struct.

We now might be able to modify the dataset_sql package to accept query objects
instead of full datasets, which is super exciting imho.

BREAKING CHANGE: the Defintion of dataset has changed. This will break hashes
from all previously run queries. So sad, but it's early days, and this is bound
to happen a bunch.