feat(Query): added new Assign and Save methods for Query/AbstractQuery
in order to do previous dataset detection, we need a way to put just
a query on a cafs to get it's hash in advance of running a query. This
beefs up Query & AbstractQuery with Assign and Save methods in the
dsfs package to make that possible without altering the query that's
passed into the save method.