feat(datatype.Json): Add Json datatype to dataset/datatype package
We've had a concept of "object" and "array" stubbed out for some
time now in the datatypes package, and it's never been used because
quite frankly the meaning of "array" and "object" is too ambiguous.
A related issue surrounds the "JsonArray" data format, which was
non-standard and silly.

So to fix that, let's remove 'Array' and 'Object' Data types in
favor of a new 'Json' data type that represents json data.

On the data format side the JsonArray data format is removed.
While the top level of all JsonDataFormat must be an array,
JsonDataFormat now defaults to object entries, and can be
switched to an array of array entries using the new
ArrayEntries FormatConfig flag in any provided structure.

This commit also brings lots of testing :)

BREAKING CHANGES: Array and Object data types are removed,
JsonArray data format is removed.