refator(Query): rename Query to Transform
For some time now I've considered SQL queries as part of a
more general category of transformations on data. The more we work
on qri, the more it seems that generalizing the dataset model opens
up more & more opportunities to integrate with other uses. This is
a first step toward supporting all kinds of other processes as on
data that also fit within this "tranformation" category.

This sets the stage for integrating with all kinds of services,
the one I'm most excited about is writing a Jupyter notebook to the
network & calling that a transform. This'll allow us to munge data
with mature tools while maintaining an auditable chain

It's worth noting that I do consider mutations that add and remove
data as a kind of transformation.

Next up is altering Transform to not be reliant on an Abstract,
supporting the case where there is not logical abstraction on the
transformation.

Fun!