refactor(dataset.Abstract): replace AbstractStructure with Abstract
In what feels like a final piece of the white paper puzzle,
abstract interoperability comparison will now be done with
abstract datasets instead of raw abstract structures.
Datasets now keep an Abstract field, which is the result of
calling Abstract on the dataset itself.

Abstract datasets are currently just their structure, abstracted,
and a dataset kind identifier. This makes it a valid dataset,
gives us the capacity to break hashes on version changes (for all
the mistakes the spec undoubtly contains).

The real win here comes in generating abstract transforms from
concrete transforms, which is now just a process of calling
Abstract on the resources, and means that there's no need for
a structural difference between abstract & concrete transforms.