test(dsfs): up test coverage of dsfs package
I've also moved a few things out of the dsfs SaveDataset method,
pushing a little more responsiblity onto the caller to arrange
the dataset properly. Data is no longer saved (it didn't make
any sense in the first place), and AbstractTranforms now
come with a check that resources are references instead
of saving them outright.

closes #31