refactor(dataset.Metadata):
ok, this one should have been obious from the start, but at least now
it's in place. In the current dataset implementation metadata would
have been duplicated all over the place, causing unacceptable space
consumption.

This refactor moves all user-added metadata into a new Metadata struct,
allowing commits that don't alter metadata to de-duplicate via hash
reference.

The byproduct of this is now Dataset is only a collection of references
which on the face of it feels right. For some time our humble dataset
package has been caught between being a child of git or a child of DCAT,
it's now clear that it's a child of git. To me this elevates the need for
interop procedures that allow datasets to be seamlessly expressed as
DCAT & datapackages, so those'll get moved up the list.