feat(dsfs.CreateDataset): initial implementaiton of dsfs.CreateDataset
This commit renames SaveDataset to WriteDataset, and adds a new
exported function on the dsfs package. CreateDataset provides a clean,
canonical implementation of how to write a dataset to a cafs.

I'm excited about this implementation b/c it's a sign of our platform
coalescing into a complete base system. It's clear to me now that
while I started this journey believing that what the world needed was
'the distribed web as a database', it now seems it's smarter to explain
what qri's becoming as a 'global data version control system', and the
first "feature" built on top of this 'dvcs' is being aware of data
transformations.

dope. a.f.

Still to do: setting structure length & row fields, data validation.