feat: rename datatypes to vals, add Value primitives
in the ongoing quest to get away from tabular structures and move
toward json as a lowest common denominator, we need to shift our
entire paradigm of data operators away from relying on the *shape*
of data in any way, and instead replace this paradigm with reliance
on data *values* being of one of seven types.

As such I've taken an initial crack by writing a vals package that
takes a great deal of inspiration from the reflect package. This'll
need lots of refinement, but should be enough to bring dsio back
from the brink of uselessness.