fix(dsfs): added set of unexported load methods
Deref methods were using their exported Load counterparts (for example
dsfs.DerefDatasetMetadata wes calling dsfs.LoadMetadata), which was causing
`/meta.json` to be appended to valid hashes of the meta file. So if hash
of the metadata file was `QmFoo...`, LoadMetadata was adding `/meta.json`,
when what we want is just `QmFoo...`. So I've added a set of unexported
methods that deref can use, sharing this code with their exported Load
counterparts.