feat(detect): added basic JSON schema detection
This is super basic, but might actually be what we want to start with. It gets the job
done, and is the minimum definable schema for a json document, which users may
appriciate, as it will cause no validation errors.

Also renamed test data files with weid "resource" suffixes to "structure"