feat(dstest): updated for go 1.10, added dstest package
first and foremost, go 1.10 has a few new complaints, cleaned those up.

Secondly, this commit brings a slew of fixes to the vals package.

testing with datasets has gotten a little bonkers. So I've started a new
packaged to help manage tests. Taking a little inspiration from the *very*
positive experiences we had using the jsonschema .json test suite, I've started
to apply a similar patten with a new package: dstest.

hopefully it's properly explained, but using dstest is a process of creating a
directory with standard filenames that have specific purposes, loading them
with `NewTestCaseFromDir`, and using the returned structure for tests. The first
place this is being put to use is in dsfs.CreateDataset.

The upside here is the capacity to use different datasets for different tests,
but present a uniform interface for working with these test cases. If we ever make it
to the point where qri needs to be implemented in a different language, these test
files will transfer over nicely.,

In the future we should think about moving the "testrepo" stuff we keep repeating
into this dataset.