fix(dsio.JSONReader): fix buffer overflow edge case reading large JSON entries
This is an interesting case that'll probably come up a bunch. Given that dataset entries can
be huge, it's possible for the size of a single JSON value to exceed the cap of bufio.Scanner's
internal buffer, causing a premature EOF.

This commit quadruples the cap of the allocated buffer to handle large tokens, at the cost of taking
up additional memory. But it's a horrible hack for the moment. It might make sense to store the
size of the largest entry in the dataset as part of a structure definition in the future to help
inform how large the cap of this buffer needs to be.