feat(CBOR): experimental support for concise binary object representation (CBOR)
We have big plans for CBOR. In the long haul we hope to make it the default data format
for qri, and to build features around CBOR as a format.

While support is unfinished, this early implementation will let us experiment with CBOR
to confirm these plans are the right direction to go, and improve our implementation using
real-world data.