refactor(dsio): alter dsio to work with Entries of primitive go types
ok, so I think the vals package might just be terrible. This replaces it in
favour of returning values with standard go types.

Stil *lots* of work to do here, but the general gist is dsio should no longer
depend on the vals package at all. `ValueReader`& `ValueWriter` are replaced
with `EntryReader` and `EntryWriter` respectively, returning a new `Entry` struct
defined in-package that just holds index, key, and value.