feat(dsio.JSONReader): rewrite JSONReader for better errors and performance
Rewrite JSONReader to use bufio.Reader instead of bufio.Scanner. Improve error handling (with test
cases), performance (x2 speedup if memory usage remains), prevent crashes if memory usage is
reduced.

Previous implementation of the JSONReader did not handle multiple types of malformed JSON. See the
test "TestJSONReaderErrors"; all test cases with error messages used to be parsed without an error
being detected.

The previous implementation of JSONReader used a huge buffer size (256KB) in order to handle large
tokens. If this new version uses the same buffer size, it improves performance by 2x. The reason is
that the old code would scan strings byte by byte, and then reparse using json.Unmarshal. This new
version only scans one, building elements as it goes. Likely, the entire runtime of the benchmarks
is dominated by the testcase "craigslist", which has an array of 12000 elements, the first of which
is approximately 10000 bytes. Scanning and copying this is going to be slow, probably slower than
all the rest of the tests.

For the previous implementation, if the buffer size were to be reduced, or if an element were
provided above the 256KB size, it would cause the JSONReader to panic. The new version fixes this
by reallocating buffers as needed, and appending them as parsing happens. Doing so is slower, but
uses less memory, and will never run into an arbitrary size limit on what can be parsed.