feat(SignableBytes): sign dataset hash & timestamp instead of title
for a while we've been signing the commit datestamp & title string.
Let's improve on that by always converting timestamps in UTC
(potentially a *very* nasty problem if left the way it was), and
signing the data hash instead of commit title.

Signing data hashes gives us the advantage of always signing a consistent
number of bytes, and being a nice concept. In this version the author is
signing off on the statement "at the point in time this data looked like this",
which I'm a fan of. the only thing I'm left wondering about is maybe adding
hash of metadata to the signature for added verification of meta info, but I'm
not totally sure on that yet. It may be equally valid to argue that metadata,
in being "data about data" isn't the principle subject of the dataset, and as such
doesn't warrent signing.