feat(detect.FromReader, dsio.TrackedReader): return number of bytes read
I'm hoping we can enforce a new rule around qri:
    _any function that accepts an io.Reader and may only read part of the reader must return the number of bytes read_
io.Reader's `Read` method returns the number of bytes read and an error. This is simply echoing that pattern. In many cases the caller of such a function will ignore this bytes-read count, but by keeping this in the return signature it both signals to the caller that "hey, this thing might do funny stuff to your reader", *and* forces the author of the function to track what they're doing to the reader. We get all of this without introducing any novel conventions, just requiring that io.Reader's return signature be honored.

With that in mind, I've started with an update to `detect.FromReader`, and added a new TrackedReader, which is just a convenice wrapper around a reader to track bytes read for readers that don't export this info (ahem, csv.Reader...)