feat(dsutil.DsYAML): added convenience func to decode from yaml
the current yaml package decodes to map[interface{}]interface{}, which isn't supported by JSON coding funcs, which creates nasty problems when you try to decode from DatasetPod (that's come from YAML) into a *dataset.Dataset. This silly little shim deals with that issue.

Plans are in place for a future version of the yaml package we use to implement decoding into map[string]interface{}. If/when that day comes we can drop this in favour of a package-native solution