fix(meta): IsEmpty function does not check to see if `Meta.License` is nil
This bug caused the dataset differ to never consider `Meta.License` when diffing the meta.

We were also marshalling and unmarshalling license to a string, not the actual license object. If this was done for a specific reason, let me know and I can change it back, but it was interferring with being able to update the license from the frontend.