feat(dataset.Body): rename Data fields to Body
After a bunch of discussion, we've decided to rename "Data" to "Body", our logic is outlined here: https://github.com/qri-io/qri/issues/422. Basically it's just a rename, but it renames a required field. This'll break all sorts of stuff, but considering we're still _very_ early days, I'd rather get it merged now.
To me this dataset-as-html-page metaphor is worth a breaking change, so let's make one.

BREAKING CHANGE: dataset.DataPath is renamed to dataset.BodyPath.