feat(viz): refactor vizconfig into initial support for render templates
@osterbit wants "qri render" templates inside of qri. After much hand-waving about security, I've thought about it, and now deeply agree that we should move to use "qri render" directly within the qri frontend to give users to express datasets however they want. So to get that to work I've done some light refactoring here, renaming to viz, and adding a slot for a single "scriptPath" file that'll be the template file.

I've removed the "Vizualizations" field that used to be in vizconfig because we don't have a use for it in this html-template direction.

The biggest change to keep an eye on is within the dsfs package, which I've adjusted to write the template file into the dataset package, following the patterns we set forth in storing transform scripts. This means the html template now travels with the dataset.