refactor(TransformResource): new Transform Reource Struct
When a transform references other datasets, it needs to record the exact path
that was used, which isn't a dataset, and we don't have the explicit concept
of a dataset reference within this package. I've introduced "TransformResource"
as a bit of a placeholder struct that should provide us room to grow in the
future, while enforcing the "path required" rule today.