refactor(meta): change 'fieldPath' -> 'fieldURL' to align with DCAT
While initially trying to get IPFS & the DCAT spec to play nice with each other
I made a decision to deviate from DCAT and rename fields like downloadURL to 'downloadPath'
to reflect the notion that either a d.web path or a proper spec-compliant URL would be used
in 'downloadPath'. I now think that's the wrong way to go, as it puts an asterisk on our
support for the DCAT spec.

This isn't a breaking change because we support arbitrary metadata. Qri will store &
preserve any previous '*Path' meta values. We'll have to update
the frontend to get and set different fields, but that won't be a big deal.