feat(pod.Assign): give Assign pattern to DatasetPod variants
turns out both plain-old-data _and_ dataset structs need the Assign pattern, for
reasons that make more sense much farther up the stack. During a refactor of our
"save path", it's important to form a unified "Pod" dataset that can be passed
in for decoding.

We still have to deal with nullifying zero values, but that'll have to come later