feat(depth): add structure depth property
this all secretly about tabular data, and properly detecting when we're dealing with it. Because JSON-schema is more about validation & description then definite requirements, we need another heuristic for checking how many levels of nesting a dataset is using at most.

We could have called this "maxDepth", but that implies there's a "minDepth", which doesn't make any sense. So, Depth.