rafactor: merge exported structs into Pod structs, remove Pod suffix
this commit drops roughly 1000 lines from the dataset package. After the cafs.Filestore interface refactor, there are no longer any prerequisite fields on a Dataset that need to be "exotic". This opens the door to making all of the major components of a dataset obey the Plain Old Data principle of only using simple types for predictable encoding & decoding. There's still lots of work to do before this is completely true.

There's still the issue of what to do about values that cannot or should-not be recorded. For this I'm introducing the phrase "transient value" for a struct field that should be removed before storing in a content-addressed file system. I'll add `RemoveTransientValues` methods next for dataset components that have them.

This is a big change, but one I think is very necessary to reduce the complexity of datasets, getting them to be easier to move around.

BREAKING CHANGE:

All dataset components exhibit the Pod pattern, pod suffix is dropped.