feat(xlsx): initial support for xlsx data format, identity dsio
in a world where we have more time I'd break these commits up. This adds initial support for the xlsx data format as readers/writers.

Also added: initial support for "identity" dsio read/write for data that's already been parsed. We're going to need this later on for working with inlined dataset bodies. currently not in use, and shouldn't be used until we have tests.

also fixed an issue with dsutil.FormFileDataset that wasn't properly assigning form values to the returned dataset.