fix(save): Fix deadlock when saving invalid file
If a user tries to save a file consisting of line-delimited json structures,
prepareDataset has the possibility of deadlocking. The reason is that
`setErrCount` and `setDepthAndEntryCount` use EntryReaders, and will only
consume the first json object. However `setChecksumAndStats` uses a normal
byte based file reader, so it will attempt to read the entire file. The
io.MultiWriter that is supplying these three goroutines will be writing
data to each of them, which will fill up the I/O buffers for `setErrCount`
and `setDepthAndEntryCount`, eventually causing `Write` within the
io.MultiWriter to block.

Fix this by closing the file handles in `setErrCount` and
`setDepthAndEntryCount` when they are done. A future improvement would be
detecting this situation somewhere and returning an error due to the trailing
content.

Fixes https://github.com/qri-io/qri/issues/660