fix(loadDatasetRefs): add correct path to dataset
In order to load the dataset from the store, we need to append the `/dataset.json` suffix to the path. However, we were incorrectly adding that same path (with the extension) to the dataset as the path.

Created another var `pathWithExt` that we use to get the dataset from the store.