refactor(writedataset): if we have a vizScript, first render the template
Because adding to the adder is destructive, we need to make sure we add to the adder in an order that makes sense.

The `dsviz.Render` process relies on the `vizScript` and `body`. Let's create the rendered file first, correctly handling the creation of TeeReaders for vizScript and Body in `dsviz.Render`.

Then we can add the rest of the files to the adder without worry!