feat(dsviz): add bodyEntries and allBodyEntries template funcs
having "paginated" access to template funcitons is important for fixed record sampling. While here I've taken the liberty of renaming "getBody" to "allBodyEntries" to clearly label for template readers that 'hey, this is going to read all the body'. With these in place we can remove the all,limit, and offset parameters from qri render.