feat(dataset): st.Strict requires validation to pass
This one is long overdue. In the future I'd like to investigate adding a more nuanced set
of validation logic to the schemas themselves that would allow progressive enhancement
of datasets (eg: "this column MUST pass validation"), but even that would be working toward
being able to set Strict: true

In the future we can chack for strict: true in packages like dsio to facilitate memory
allocation speedups. User interfaces can show some sort of check or seal declaring the
dataset as "clean data". Finally, we should be able to do cool things when two datasets are
strict, thanks to the fact that we know the data won't contain any surprises.