feat(ShallowCompare,PathMap): add utility methods for comparing components
ShallowCompare is useful for comparing components that are in different states
of persistence. For example one "viz" component is saved to /ipfs/, and the
other "viz" is not yet saved. Or one "viz" component is saved to /s3/ and the
other is saved to /ipfs/.

Also add tests for readme component.