feat(dataset): add ID field to dataset.Dataset
in qri-io/qri land we talk about a cascade of values from refstring to dataset,
where each structure is a superset of the structure before it:

  * reference string: "b5/population@/ipfs/hash"
  * dsref.Ref: { "username": "b5", "name": "population", "path": "/ipfs/hash" }
  * dsref.VersionInfo: { "username": "b5", "name": "population", "path": "/ipfs/hash", "commitTitle": "added..." }
  * dataset.Dataset: { /* complete dataset */ }

This lets us expand-to-suit depending on the needs of the operation. It also
means that any field that can be represented in a smaller structure must be
present in all others.

In a separate converation we've been on a push to use _stable identifiers for
dataset histories_, which have to be worked into _all_ of these structures to be
effective. This commit adds an ID field to the dataset structure itself to
advance that goal. the ID field is added directly to the dataset struct, should
be persisted to disk, and is added to the `SignableBytes` method to add
cryptographic integrity to it's value.

There is already an "ID" field on a dataset within the Meta component. We can't
use it for this purpose and keep the `Meta` component optional, so I've chosen
to add it to the dataset itself. I generally think of `Meta.ID` as being a
refernce to an _external_ identifier.

BREAKING CHANGE:
older versions of qri that attempt to verify the signature of datasets with a
non-empty ID string field will error.